+++ include/sparrow/c_interface/arrow_schema/deleter.hpp
@@ -1,0 +1,43 @@
+// Copyright 2024 Man Group Operations Limited
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "sparrow/c_interface/c_structures.hpp"
+
+namespace sparrow
+{
+    inline void arrow_schema_custom_deleter(ArrowSchema* schema)
+    {
+        if (schema)
+        {
+            if (schema->release != nullptr)
+            {
+                schema->release(schema);
+            }
+            delete schema;
+        }
+    }
+
+    /**
+     * A custom deleter for ArrowSchema.
+     */
+    struct arrow_schema_custom_deleter_struct
+    {
+        void operator()(ArrowSchema* schema) const
+        {
+            arrow_schema_custom_deleter(schema);
+        }
+    };
+}
