+++ test/test_arrow_array_schema_utils.cpp
@@ -14,7 +14,7 @@
 
 #include <cstddef>
 
-#include "sparrow/arrow_array_schema_utils.hpp"
+#include "sparrow/c_interface/arrow_array_schema_utils.hpp"
 #include "sparrow/buffer.hpp"
 
 #include "doctest/doctest.h"
